requirements.txt
# Tigo Honduras Multimodal RAG System - Requirements

# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.0

# HTTP requests
requests==2.31.0

# Azure services
openai>=1.0.0
azure-search-documents>=11.4.0
azure-core>=1.29.0
azure-identity>=1.15.0

# Data processing and ML
numpy==1.24.3
scikit-learn==1.3.0
pandas>=1.5.0
openpyxl>=3.0.0

# Image processing
Pillow==10.1.0

# JSON and data handling
python-json-logger==2.0.7

# Environment and configuration
python-dotenv==1.0.0

# Type hints and development
typing-extensions==4.8.0

# Additional dependencies for production
gunicorn>=21.2.0
aiofiles>=23.2.1

# Audio processing (optional but recommended)
# librosa==0.10.1
# soundfile==0.12.1